requirements.txt
# Flexible requirements.txt

# Core FastAPI dependencies
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
starlette>=0.27.0,<0.28.0
pydantic>=2.5.0,<2.6.0
pydantic-settings>=2.0.3,<2.1.0
itsdangerous>=2.1.2,<2.2.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.6,<0.1.0
bcrypt>=4.1.0,<4.2.0
cryptography>=41.0.0,<42.0.0  # Flexible version range

# Database
sqlalchemy>=2.0.20,<2.1.0
psycopg2-binary>=2.9.7,<2.10.0
alembic>=1.12.0,<1.13.0

# Template and Static Files
jinja2>=3.1.0,<3.2.0
aiofiles>=23.2.0,<24.0.0

# HTTP Requests and Image Processing  
requests>=2.31.0,<2.32.0
pillow>=10.0.0,<11.0.0
httpx>=0.25.0,<0.26.0

# Environment and Configuration
python-dotenv>=1.0.0,<1.1.0

# Additional utilities
python-dateutil>=2.8.0,<2.9.0


